[@svelteui/core]: rename utils folder
