+++ packages/svelteui-core/src/components/Form/utils/lists/change-error-indices.ts
