+++ packages/svelteui-core/src/components/Form/utils/paths/get-path.ts
