+++ packages/svelteui-core/src/components/Form/utils/paths/get-splitted-path.ts
