+++ packages/svelteui-core/src/components/Form/utils/paths/remove-path.ts
