+++ packages/svelteui-core/src/components/Form/utils/paths/reorder-path.ts
