+++ packages/svelteui-core/src/components/Form/utils/validators/is-email/is-email.ts
