+++ packages/svelteui-core/src/components/Form/utils/validators/is-not-empty/is-not-empty.ts
